+++ src/InsecureSigningMethods.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Penneo\SDK;
+
+class InsecureSigningMethods
+{
+    public const DRAW = 'draw';
+    public const IMAGE = 'image';
+    public const TEXT = 'text';
+}