+++ src/SigningRequest.php
@@ -2,7 +2,7 @@
 
 namespace Penneo\SDK;
 
-use Penneo\SDK\Entity;
+use InvalidArgumentException;
 
 class SigningRequest extends Entity
 {
@@ -20,7 +20,8 @@ class SigningRequest extends Entity
             'failUrl',
             'reminderInterval',
             'accessControl',
-            'enableInsecureSigning'
+            'enableInsecureSigning',
+            'insecureSigningMethods',
         )
     );
     protected static $relativeUrl = 'signingrequests';
@@ -40,6 +41,14 @@ class SigningRequest extends Entity
     protected $reminderInterval;
     protected $accessControl;
     protected $enableInsecureSigning;
+
+    /**
+     * The simple electronic signing methods that the signer
+     * can use to make a simple electronic signature
+     *
+     * @var array<"draw" | "image" | "text">|null $insecureSigningMethods
+     */
+    protected $insecureSigningMethods;
 
     public function getLink()
     {
@@ -205,5 +214,40 @@ class SigningRequest extends Entity
     public function setEnableInsecureSigning($enableInsecureSigning)
     {
         $this->enableInsecureSigning = $enableInsecureSigning;
+    }
+
+    public function getInsecureSigningMethods(): ?array
+    {
+        return $this->insecureSigningMethods;
+    }
+
+    /**
+     * See InsecureSigningMethods class for possible values
+     *
+     * @param  array<"draw" | "image" | "text">|null $insecureSigningMethods
+     * @return void
+     */
+    public function setInsecureSigningMethods(?array $insecureSigningMethods): void
+    {
+        if ($insecureSigningMethods !== null) {
+            foreach ($insecureSigningMethods as $method) {
+                if (
+                    !in_array(
+                        $method,
+                        [
+                        InsecureSigningMethods::DRAW,
+                        InsecureSigningMethods::IMAGE,
+                        InsecureSigningMethods::TEXT],
+                        true
+                    )
+                ) {
+                    throw new InvalidArgumentException(
+                        'Invalid signing method: ' . $method
+                    );
+                }
+            }
+        }
+
+        $this->insecureSigningMethods = $insecureSigningMethods;
     }
 }